pyproject.toml
[project]
name = "mba-deobfuscator"
version = "0.1.0"
description = "GNN+Transformer system for simplifying obfuscated Mixed Boolean-Arithmetic expressions"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10,<3.13"
authors = [
    {name = "MBA Deobfuscator Team"}
]
keywords = ["machine-learning", "deobfuscation", "mba", "gnn", "transformer"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Core dependencies (torch/torch-geometric installed separately via setup.sh)
dependencies = [
    "numpy>=1.24.0",
    "z3-solver>=4.12.0",
    "tensorboard>=2.13.0",
    "wandb>=0.15.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.scripts]
mba-train = "scripts.train:main"
mba-eval = "scripts.evaluate:main"
mba-simplify = "scripts.simplify:main"

[project.urls]
Repository = "https://github.com/your-org/mba-deobfuscator"
Documentation = "https://github.com/your-org/mba-deobfuscator/tree/main/docs"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | build
    | dist
    | __pycache__
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = [".git", ".venv", "build", "dist"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]
